demo for issue 635
however, I have realized my error.
I was follow the epic stack in my remix app and was not ready for Sentry. Therefore the comment about it in `server/index.ts` made me this I could get rid of the NODE_ENV check.
Therefore, I was left with just the `await getBuild()` call.
However, because I was calling `getBuild` now in development the `viteDevServer.ssrLoadModule` part needed to be wrapped in a function like the remix/website example

To summarize. If you try this branch out and remove the wrapper function around `viteDevServer.ssrLoadModule` you will see that the app starts but updates inside a loader are not reflected. Which is where I thought to help